+++ includes/class-setup-plugin.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.0
+ * @version  1.0.1
  */
 class PfBT_Setup_Plugin {
 
@@ -14,13 +14,31 @@ class PfBT_Setup_Plugin {
 	 * Initialization.
 	 *
 	 * @since    1.0.0
-	 * @version  1.0.0
+	 * @version  1.0.1
 	 */
 	public static function init() {
 		add_action( 'pfbt_content', __CLASS__ . '::notice' );
+		add_action( 'pfbt_content', __CLASS__ . '::content' );
 
 		add_filter( 'fl_theme_builder_part_hooks', __CLASS__ . '::parts' );
 		add_filter( 'fl_builder_upgrade_url', __CLASS__ . '::url' );
+	}
+
+	/**
+	 * Is builder active or enabled?
+	 *
+	 * @since    1.0.1
+	 * @version  1.0.1
+	 */
+	public static function is_builder() {
+		if (
+			! is_callable( 'FLBuilderModel::is_builder_active' )
+			|| ! is_callable( 'FLBuilderModel::is_builder_enabled' )
+		) {
+			return false;
+		}
+
+		return FLBuilderModel::is_builder_active() || FLBuilderModel::is_builder_enabled();
 	}
 
 	/**
@@ -30,18 +48,31 @@ class PfBT_Setup_Plugin {
 	 * but there is no Beaver Themer layout to display.
 	 *
 	 * @since    1.0.0
-	 * @version  1.0.0
+	 * @version  1.0.1
 	 */
 	public static function notice() {
 		if ( ! class_exists( 'FLThemeBuilder' ) ) {
 			// Beaver Themer not active?
 			get_template_part( 'templates/parts/content/content', 'beaver-themer' );
-		} else if (
-			! is_callable( 'FLBuilderModel::is_builder_active' )
-			|| ! FLBuilderModel::is_builder_active()
-		) {
+		} else if ( ! self::is_builder() ) {
 			// Looks like we have no Themer Layouts...
 			get_template_part( 'templates/parts/content/content', 'themer-layouts' );
+		}
+	}
+
+	/**
+	 * Display page builder layout if it exists.
+	 *
+	 * Allows displaying custom page builder layout if it exists,
+	 * such as pages built with Beaver Builder. This will bypass
+	 * the Beaver Themer layout.
+	 *
+	 * @since    1.0.1
+	 * @version  1.0.1
+	 */
+	public static function content() {
+		if ( self::is_builder() ) {
+			get_template_part( 'templates/parts/content/content', get_post_type() );
 		}
 	}