+++ changelog.md
@@ -1,5 +1,22 @@
 # Playground for Beaver Themer Changelog
 
+## 1.0.1
+
+* **Fix**: CSS font family values
+* **Fix**: Making the theme work with Beaver Builder pages/posts
+
+### Files changed:
+
+    changelog.md
+    index.php
+    style.css
+    assets/css/base.css
+    assets/sass/base.scss
+    assets/sass/parts/_typography.scss
+    includes/class-setup-plugin.php
+    templates/parts/content/content.php
+
+
 ## 1.0.0
 
 * Initial release.